[CLEANUP] signedness fixes in some compare operations
At some places, we compared signed with unsigned, which makes the compiler
complain. Fix that either by using explicit signed ints or by casting during
the compare.

+++ eb32tree.h
@@ -184,7 +184,7 @@ static forceinline struct eb32_node *__eb32i_lookup(struct eb_root *root, s32 x)
 		if ((eb_gettag(troot) == EB_LEAF)) {
 			node = container_of(eb_untag(troot, EB_LEAF),
 					    struct eb32_node, node.branches);
-			if (node->key == x)
+			if (node->key == (u32)x)
 				return node;
 			else
 				return NULL;
